poparser: avoid comparisons of different signs

+++ src/poparser.c
@@ -162,7 +162,7 @@ enum po_error poparser_feed_line(struct po_parser *p, char* in, size_t in_len) {
 			line_len = len + 1;
 			x = p->buf;
 			len = p->bufsize;
-			if ( (len = iconv(p->cd, &line, &line_len, &x, &len)) == -1) 
+			if ( (len = iconv(p->cd, &line, &line_len, &x, &len)) == (size_t)-1) 
 				return -po_failed_iconv;
 
 			x = p->buf;
@@ -222,7 +222,7 @@ enum po_error poparser_feed_line(struct po_parser *p, char* in, size_t in_len) {
 			line_len = len + 1;
 			x = p->buf;
 			len = p->bufsize;
-			if ( (len = iconv(p->cd, &line, &line_len, &x, &len)) == -1) 
+			if ( (len = iconv(p->cd, &line, &line_len, &x, &len)) == (size_t)-1) 
 				return -po_failed_iconv;
 
 			x = p->buf;
